requirements.txt
pyaml
python-dotenv